+++ src/Sonar/Aware/ModelInterface.php
@@ -18,7 +18,6 @@ interface ModelInterface {
      * Receiving incoming data
      *
      * @param mixed $data
-     * @return void
      */
     public function __construct($data);