+++ src/Sonar/Services/CacheService.php
@@ -1,10 +1,10 @@
 <?php
 namespace Sonar\Services;
 
-use Ratchet\Session\SessionProvider;
-use Ratchet\MessageComponentInterface;
-use Symfony\Component\HttpFoundation\Session\Storage\Handler\MemcacheSessionHandler;
-
+use Sonar\Exceptions\CacheServiceException;
+use Phalcon\Cache\Backend\Memcache;
+use Phalcon\Cache\Frontend\Data;
+use Phalcon\Cache\Exception as CacheException;
 /**
  * Class CacheService. Cache data service
  *
@@ -21,38 +21,39 @@ class CacheService {
     /**
      * Memcache connect
      *
-     * @var \Memcache $cache
+     * @var \Phalcon\Cache\Backend\Memcache $connect
      */
-    private $cache;
-
-    /**
-     * Session provider instance
-     *
-     * @var \Ratchet\Session\SessionProvider $provider
-     */
-    private $provider;
+    private $connect;
 
     /**
      * Implement chache configurations
      */
     public function __construct(\Phalcon\Config $config) {
 
-        $this->cache = new \Memcache();
-        $this->cache->connect($config->host, $config->port);
+        try {
+
+            if(!$this->connect) {
+
+                $this->connect = new Memcache(new Data(['lifetime' => $config->lifetime]), [
+                    'host'      =>  $config->host,
+                    'port'      =>  $config->port,
+                    'prefix'    =>  $config->prefix,
+                    'persistent'=>  $config->persistent,
+                ]);
+
+            }
+        }
+        catch(CacheException $e) {
+            throw new CacheServiceException($e->getMessage());
+        }
     }
 
     /**
-     * Add application to cache storage
+     * Return cache storage instance
      *
-     * @param \MessageComponentInterface $app
-     * @return \Ratchet\Session\SessionProvider
+     * @return \Phalcon\Cache\Backend\Memcache
      */
-    public function addApplication(\MessageComponentInterface $app) {
-
-        if(!$this->provider) {
-            $this->provider = new SessionProvider($app,  new MemcacheSessionHandler($this->cache));
-        }
-
-        return $this->provider;
+    public function getStorage() {
+        return $this->connect;
     }
 }